requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
asyncer==0.0.5
attrs==23.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
boto3==1.34.25
botocore==1.34.25
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
dj-database-url==0.5.0
Django==4.0.2
django-apscheduler==0.6.2
django-cors-headers==4.3.1
djangorestframework==3.13.1
easyocr==1.6.2
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.7
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gradio==4.21.0
gradio_client==0.12.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.3
imageio==2.25.0
importlib_metadata==7.0.2
importlib_resources==6.1.3
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
#mysql-client==0.0.1
networkx==3.0
ninja==1.11.1
numba==0.58.1
numpy==2.2.4
onnxruntime==1.16.3
opencv-python-headless==4.5.4.60
orjson==3.9.15
packaging==23.0
pandas==2.0.3
Pillow==9.0.1
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
pyclipper==1.3.0.post4
pycodestyle==2.8.0
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
PyMatting==1.1.12
PyMySQL==1.0.2
pyparsing==3.1.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-multipart==0.0.9
pytz==2021.3
PyWavelets==1.4.1
PyYAML==6.0
referencing==0.33.0
rembg==2.0.65
requests==2.28.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.2
s3transfer==0.10.0
scikit-image==0.19.3
scipy==1.10.0
semantic-version==2.10.0
shapely==2.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.2
starlette==0.36.3
sympy==1.12
tifffile==2023.1.23.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
Unipath==1.1
urllib3==1.26.9
uvicorn==0.28.0
uWSGI==2.0.21
watchdog==4.0.0
websockets==11.0.3
whitenoise==5.3.0
yarl==1.9.4
zipp==3.17.0
